Add 'useContext' for state management
