+++ src/App.js
@@ -1,9 +1,6 @@
-import { useState } from 'react';
-import { v4 as uuidv4 } from 'uuid';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Header from './components/Header'
 import FeedbackList from './components/FeedbackList';
-import FeedbackData from './data/FeedbackData';
 import FeedbackStats from './components/FeedbackStats';
 import FeedbackForm from './components/FeedbackForm';
 import AboutIconLink from './components/AboutIconLink';
@@ -11,21 +8,6 @@ import AboutPage from './pages/AboutPage';
 import { FeedbackProvider } from './context/FeedbackContext'
 
 const App = () => {
-  const [feedback, setFeedback] = useState(FeedbackData)
-
-  const addFeedback = (newFeedback) => {
-    newFeedback.id = uuidv4()
-    setFeedback([newFeedback, ...feedback])
-  }
-
-  const deleteFeedback = (id) => {
-    if(window.confirm("This will delete your feedback.")) {
-      setFeedback((prevState) => (
-        prevState.filter((feedback) => feedback.id !== id)
-      ))
-    }
-  }
-
   return (
     <FeedbackProvider>
       <Router>
@@ -37,9 +19,9 @@ const App = () => {
               path="/"
               element={
                 <>
-                  <FeedbackForm handleAdd={addFeedback} />
-                  <FeedbackStats feedback={feedback} />
-                  <FeedbackList feedback={feedback} handleDelete={deleteFeedback} />
+                  <FeedbackForm />
+                  <FeedbackStats />
+                  <FeedbackList />
                 </>
               }
             ></Route>
