+++ src/components/FeedbackForm.jsx
@@ -1,13 +1,16 @@
-import { useState } from "react"
+import { useState, useContext } from "react"
 import Card from "./shared/Card"
 import Button from "./shared/Button"
 import RatingSelect from "./shared/RatingSelect"
+import FeedbackContext from "../context/FeedbackContext"
 
-const FeedbackForm = ({ handleAdd }) => {
+const FeedbackForm = () => {
   const [text, setText] = useState('')
   const [rating, setRating] = useState(10)
   const [btnDisabled, setBtnDisabled] = useState(true)
   const [validateMsg, setValidateMsg] = useState('')
+
+  const { addFeedback } = useContext(FeedbackContext)
 
   const handleTextChange = ({ target: { value } }) => {
     if (value === '') {
@@ -32,7 +35,8 @@ const FeedbackForm = ({ handleAdd }) => {
         rating,
       }
 
-      handleAdd(newFeedback)
+      addFeedback(newFeedback)
+
       setText('')
     }
   }
