+++ src/components/FeedbackItem.jsx
@@ -1,22 +1,20 @@
 import { FaTimes } from 'react-icons/fa'
-import PropTypes from 'prop-types'
+import { useContext } from 'react'
 import Card from "./shared/Card"
+import FeedbackContext from '../context/FeedbackContext'
 
-const FeedbackItem = ({ item, handleDelete }) => {
+const FeedbackItem = ({ item }) => {
+  const { deleteFeedback } = useContext(FeedbackContext)
+
   return (
     <Card>
       <div className="num-display">{item.rating}</div>
-      <button onClick={() => handleDelete(item.id)} className="close">
+      <button onClick={() => deleteFeedback(item.id)} className="close">
         <FaTimes color="purple" />
       </button>
       <div className="text-display">{item.text}</div>
     </Card>
   )
-}
-
-FeedbackItem.propTypes = {
-  item: PropTypes.object.isRequired,
-  handleDelete: PropTypes.func
 }
 
 export default FeedbackItem
