+++ src/components/FeedbackList.jsx
@@ -1,8 +1,11 @@
 import { motion, AnimatePresence } from 'framer-motion';
-import PropTypes from 'prop-types'
+import { useContext } from 'react'
 import FeedbackItem from './FeedbackItem';
+import FeedbackContext from '../context/FeedbackContext';
 
-const FeedbackList = ({ feedback, handleDelete }) => {
+const FeedbackList = () => {
+  const {feedback} = useContext(FeedbackContext)
+
   if (!feedback || feedback.length === 0) {
     return <p>No Feedback Yet.</p>
   }
@@ -20,24 +23,12 @@ const FeedbackList = ({ feedback, handleDelete }) => {
             <FeedbackItem 
               key={item.id}
               item={item}
-              handleDelete={handleDelete}
             />
           </motion.div>
         ))}
       </AnimatePresence>
     </div>
   )
-}
-
-FeedbackList.propTypes = {
-  feedback: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.number.isRequired,
-      text: PropTypes.string.isRequired,
-      rating: PropTypes.number.isRequired,
-    })
-  ),
-  handleDelete: PropTypes.func,
 }
 
 export default FeedbackList
