+++ src/context/FeedbackContext.js
@@ -1,3 +1,4 @@
+import { v4 as uuidv4 } from 'uuid'
 import { createContext, useState } from 'react'
 
 const FeedbackContext = createContext()
@@ -6,14 +7,41 @@ export const FeedbackProvider = ({children}) => {
   const [feedback, setFeedback] = useState([
     {
       id: 1,
-      text: 'This item is from context',
+      text: 'This is feedback item one(1)',
       rating: 10
-    }
+    },
+    {
+      id: 2,
+      text: 'This is feedback item two(2)',
+      rating: 9
+    },
+    {
+      id: 3,
+      text: 'This is feedback item three(3)',
+      rating: 7
+    },
   ])
+
+  const addFeedback = (newFeedback) => {
+    newFeedback.id = uuidv4()
+    setFeedback([newFeedback, ...feedback])
+  }
+
+  const deleteFeedback = (id) => {
+    if(window.confirm("This will delete your feedback.")) {
+      setFeedback((prevState) => (
+        prevState.filter((feedback) => feedback.id !== id)
+      ))
+    }
+  }
 
   return (
     <FeedbackContext.Provider 
-      value={{ feedback }}
+      value={{ 
+        feedback,
+        addFeedback,
+        deleteFeedback,
+      }}
     >
       {children}
     </FeedbackContext.Provider>
